requirements.txt
protobuf==5.26.0
streamlit==1.32.2
google-generativeai
